+++ pkg/models/settings.go
@@ -3,8 +3,10 @@ package models
 import (
 	"context"
 	"encoding/json"
+	"errors"
 	"fmt"
 	"net/textproto"
+	urllib "net/url"
 	"strings"
 
 	"github.com/grafana/grafana-plugin-sdk-go/backend"
@@ -44,6 +46,8 @@ type OAuth2Settings struct {
 	Subject        string           `json:"subject,omitempty"`
 	Scopes         []string         `json:"scopes,omitempty"`
 	AuthStyle      oauth2.AuthStyle `json:"authStyle,omitempty"`
+	AuthHeader     string           `json:"authHeader,omitempty"`
+	TokenTemplate  string           `json:"tokenTemplate,omitempty"`
 	ClientSecret   string
 	PrivateKey     string
 	EndpointParams map[string]string
@@ -119,6 +123,7 @@ type InfinitySettings struct {
 	AzureBlobAccountKey       string
 	UnsecuredQueryHandling    UnsecuredQueryHandlingMode
 	PathEncodedURLsEnabled    bool
+	IgnoreStatusCodeCheck     bool
 	AllowDangerousHTTPMethods bool
 	// ProxyOpts is used for Secure Socks Proxy configuration
 	ProxyOpts httpclient.Options
@@ -154,19 +159,26 @@ func (s *InfinitySettings) Validate() error {
 		}
 		return nil
 	}
-	if s.AuthenticationMethod != AuthenticationMethodAzureBlob && s.AuthenticationMethod != AuthenticationMethodNone && len(s.AllowedHosts) < 1 {
+	if s.DoesAllowedHostsRequired() && len(s.AllowedHosts) < 1 {
 		return ErrInvalidConfigHostNotAllowed
 	}
-	if s.HaveSecureHeaders() && len(s.AllowedHosts) < 1 {
-		return ErrInvalidConfigHostNotAllowed
-	}
-	if len(s.KeepCookies) > 0 && len(s.AllowedHosts) < 1 {
-		return ErrInvalidConfigHostNotAllowed
-	}
-	return nil
+	return ValidateAllowedHosts(s.AllowedHosts)
 }
 
-func (s *InfinitySettings) HaveSecureHeaders() bool {
+func (s *InfinitySettings) DoesAllowedHostsRequired() bool {
+	// If base url is configured, there is no need for allowed hosts
+	if strings.TrimSpace(s.URL) != "" {
+		return false
+	}
+	// If there is specific authentication mechanism (except none and azure blob), then allowed hosts required
+	if s.AuthenticationMethod != "" && s.AuthenticationMethod != AuthenticationMethodNone && s.AuthenticationMethod != AuthenticationMethodAzureBlob {
+		return true
+	}
+	// If there are any TLS specific settings enabled, then allowed hosts required
+	if s.TLSAuthWithCACert || s.TLSClientAuth {
+		return true
+	}
+	// If there are custom headers (not generic headers such as Accept, Content Type etc), then allowed hosts required
 	if len(s.CustomHeaders) > 0 {
 		for k := range s.CustomHeaders {
 			if textproto.CanonicalMIMEHeaderKey(k) == "Accept" {
@@ -177,9 +189,49 @@ func (s *InfinitySettings) HaveSecureHeaders() bool {
 			}
 			return true
 		}
-		return false
+	}
+	// If there are custom query parameters, then allowed hosts required
+	if len(s.SecureQueryFields) > 0 {
+		return true
+	}
+	// If there are cookie forwarding, then allowed hosts required
+	if len(s.KeepCookies) > 0 {
+		return true
 	}
 	return false
+}
+
+func ValidateAllowedHosts(allowedUrls []string) error {
+	for _, allowedUrl := range allowedUrls {
+		fullAllowedUrl := FixMissingURLSchema(allowedUrl)
+		parsedURL, err := urllib.Parse(fullAllowedUrl)
+		if err != nil {
+			return fmt.Errorf("error parsing allowed list url %s", allowedUrl)
+		}
+		allowedHostName := parsedURL.Hostname()
+		if strings.TrimSpace(allowedHostName) == "" {
+			return errors.New("invalid url found in allowed hosts settings")
+		}
+		protocol := strings.ToLower(parsedURL.Scheme)
+		if !(protocol == "http" || protocol == "https") {
+			return fmt.Errorf("invalid url in allowed list %s", allowedUrl)
+		}
+	}
+	return nil
+}
+
+func FixMissingURLSchema(urlString string) string {
+	if strings.TrimSpace(urlString) == "" {
+		return ""
+	}
+	lowerUrlString := strings.ToLower(urlString)
+	if !(strings.HasPrefix(lowerUrlString, "http://") || strings.HasPrefix(lowerUrlString, "https://")) {
+		if lowerUrlString == "localhost" || strings.HasPrefix(lowerUrlString, "localhost:") {
+			return "http://" + urlString
+		}
+		return "https://" + urlString
+	}
+	return urlString
 }
 
 type RefData struct {
@@ -210,6 +262,7 @@ type InfinitySettingsJson struct {
 	AzureBlobAccountUrl       string         `json:"azureBlobAccountUrl,omitempty"`
 	AzureBlobAccountName      string         `json:"azureBlobAccountName,omitempty"`
 	PathEncodedURLsEnabled    bool           `json:"pathEncodedUrlsEnabled,omitempty"`
+	IgnoreStatusCodeCheck     bool           `json:"ignoreStatusCodeCheck,omitempty"`
 	AllowDangerousHTTPMethods bool           `json:"allowDangerousHTTPMethods,omitempty"`
 	// Security
 	AllowedHosts           []string                   `json:"allowedHosts,omitempty"`
@@ -256,6 +309,7 @@ func LoadSettings(ctx context.Context, config backend.DataSourceInstanceSettings
 		settings.ProxyUrl = infJson.ProxyUrl
 		settings.ProxyUserName = infJson.ProxyUserName
 		settings.PathEncodedURLsEnabled = infJson.PathEncodedURLsEnabled
+		settings.IgnoreStatusCodeCheck = infJson.IgnoreStatusCodeCheck
 		settings.AllowDangerousHTTPMethods = infJson.AllowDangerousHTTPMethods
 		if settings.ProxyType == "" {
 			settings.ProxyType = ProxyTypeEnv